refactor: add the Player and the Game classes
