+++ utils.js
@@ -1,2 +1,19 @@
 //Randomized HP function
 export const getRandom = (num) => Math.ceil(Math.random() * num);
+
+// Create player DOM
+function createPlayer(playerClass, player) {
+  return `
+  <div class="${playerClass}">
+    <div class="progressbar">
+        <div class="life" style="width: ${player.hp}%;"></div>
+        <div class="name">${player.name}</div>
+    </div>
+    <div class="character">
+        <img src="${player.img}" />
+    </div>
+  </div>
+  `;
+}
+
+export default createPlayer;
